Move functions from llpcShaderModuleHelper to vkgcUtil
Move isSpirvBinary() and getEntryPointNameFromSpirvBinary() from llpcShaderModuleHelper to vkgcUtil to make them compiler independent components.
